requirements.txt
tensorflow
keras==2.15.0
numpy
pillow
matplotlib
tqdm
transformers
torch
streamlit